[updated] tool added world snapshot table
git-svn-id: http://svn2.assembla.com/svn/swgemu/tools@3548 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ WorldSpawnerTool/external/meshLib/src/eft.cpp
@@ -183,7 +183,7 @@ unsigned int eft::readSCAP( std::istream &file )
     for( unsigned int i = 0; i < numScap; ++i )
     {
       total += base::read( file, data );
-      std::cout << std::bitset<32>( data ) << " " << std::endl;
+      std::cout << std::bitset<32>( (int)data ) << " " << std::endl;
     }
     total += scapSize;
